Extract JWT token handling to cactus-common library
